+++ app/Models/Ticket.php
@@ -57,11 +57,13 @@ class Ticket extends Init
         parent::boot();
 
         static::deleting(function (Model $model) {
-            $file = 'images/' . $model->image->uri ?? '';
-            if (Storage::disk('public')->exists($file)) {
-                Storage::disk('public')->delete($file);
+            if (!empty($model->image?->uri)) {
+                $file = 'images/' . $model->image->uri ?? '';
+                if (Storage::disk('public')->exists($file)) {
+                    Storage::disk('public')->delete($file);
+                }
+                $model->image()->delete();
             }
-            $model->image()->delete();
         });
     }
 }
